backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
cryptography==41.0.7
alembic==1.13.0
python-dotenv==1.0.0
bcrypt==4.1.1
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
user-agents==2.2.0